Fix more warnings

See merge request GNOME/glib!2074

+++ tests/gobject/timeloop-closure.c
@@ -33,7 +33,7 @@ io_pipe (GIOChannel **channels)
 }
 
 static gboolean
-read_all (GIOChannel *channel, char *buf, int len)
+read_all (GIOChannel *channel, char *buf, gsize len)
 {
   gsize bytes_read = 0;
   gsize count;
@@ -57,7 +57,7 @@ read_all (GIOChannel *channel, char *buf, int len)
 }
 
 static gboolean
-write_all (GIOChannel *channel, char *buf, int len)
+write_all (GIOChannel *channel, char *buf, gsize len)
 {
   gsize bytes_written = 0;
   gsize count;
